renamed example files to snake_case format
