+++ examples/panda_reacher.py
@@ -16,15 +16,15 @@ def main():
     ob = env.reset()
     print(f"Initial observation : {ob}")
     if goal:
-        from examples.sceneObjects.goal import dynamicGoal
+        from examples.scene_objects.goal import dynamicGoal
         env.add_goal(dynamicGoal)
 
     if obstacles:
-        from examples.sceneObjects.obstacles import dynamicSphereObst2
+        from examples.scene_objects.obstacles import dynamicSphereObst2
         env.add_goal(dynamicGoal)
 
     if obstacles:
-        from examples.sceneObjects.obstacles import dynamicSphereObst2
+        from examples.scene_objects.obstacles import dynamicSphereObst2
 
         env.add_obstacle(dynamicSphereObst2)
     print("Starting episode")
