+++ examples/scene_objects/goal.py
