+++ examples/scene_objects/obstacle_data/duck.obj
