+++ examples/scene_objects/obstacle_data/duck_vhacd.obj
