+++ examples/scene_objects/obstacle_data/sphere_goal.urdf
