+++ examples/scene_objects/obstacle_data/sphere_smooth.obj
