+++ urdfenvs/sensors/obstacle_sensor.py
@@ -16,7 +16,7 @@ class ObstacleSensor(Sensor):
 
     _observation: dict
         For every object the Pose and Twist are stored in the observation.
-        Pose contains position in cartesian format (x, y, z) and orientation in quaternion format (a, b, c, d)
+        Pose contains position in cartesian format (x, y, z) and orientation in quaternion format (x, y, z, w)
         Twist contains velocity in cartesian format (x, y, z) and angular velocity in cartesian format (x, y, z)
 
     """
